task/security - last changes and minor improvements

+++ src/main/java/com/vention/stockmarket/service/impl/FavouriteCompaniesServiceImpl.java
@@ -1,11 +1,12 @@
 package com.vention.stockmarket.service.impl;
 
 import com.vention.stockmarket.dto.response.FavouriteCompanyDTO;
+import com.vention.stockmarket.exceptions.CustomUnauthorizedException;
 import com.vention.stockmarket.repository.FavouriteCompaniesRepository;
 import com.vention.stockmarket.repository.SecurityRepository;
 import com.vention.stockmarket.repository.StockRepository;
-import com.vention.stockmarket.repository.UserRepository;
 import com.vention.stockmarket.service.FavouriteCompaniesService;
+import com.vention.stockmarket.service.SecurityHelperService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -18,8 +19,12 @@ public class FavouriteCompaniesServiceImpl implements FavouriteCompaniesService 
     private final FavouriteCompaniesRepository repository;
     private final SecurityRepository securityRepository;
     private final StockRepository stockRepository;
+
     @Override
     public void addToFavorites(String username, String companyName) {
+        if (!SecurityHelperService.hasUserPermissions(username)) {
+            throw new CustomUnauthorizedException();
+        }
         var byEmail = securityRepository.getByEmail(username);
         var bySymbol = stockRepository.findBySymbol(companyName).get();
         repository.add(byEmail.get().getUserId(), bySymbol.getSymbol());
@@ -27,6 +32,9 @@ public class FavouriteCompaniesServiceImpl implements FavouriteCompaniesService 
 
     @Override
     public List<FavouriteCompanyDTO> getAllByUsername(String username) {
+        if (!SecurityHelperService.hasUserPermissions(username)) {
+            throw new CustomUnauthorizedException();
+        }
         var byEmail = securityRepository.getByEmail(username);
         var companiesSymbols = repository.findByUserId(byEmail.get().getUserId());
         var all = stockRepository.findAll(companiesSymbols);
@@ -36,6 +44,9 @@ public class FavouriteCompaniesServiceImpl implements FavouriteCompaniesService 
 
     @Override
     public void delete(String username, String symbol) {
+        if (!SecurityHelperService.hasUserPermissions(username)) {
+            throw new CustomUnauthorizedException();
+        }
         var byEmail = securityRepository.getByEmail(username);
         var bySymbol = stockRepository.findBySymbol(symbol).get();
         repository.delete(byEmail.get().getUserId(), bySymbol.getSymbol());
